+++ src/components/BindPhraseInput.vue
@@ -1,0 +1,21 @@
+<script setup>
+import {ref} from "vue";
+import {VTextField} from "vuetify/components";
+import {uidBytesFromText} from "../phrase.js";
+
+let model = defineModel()
+
+let bindPhrase = ref(null)
+let uid=ref('Bind Phrase')
+function generateUID() {
+  if (bindPhrase.value === '') uid.value='Bind Phrase'
+  else {
+    model = uidBytesFromText(bindPhrase.value)
+    uid.value = 'UID: ' + model
+  }
+}
+</script>
+
+<template>
+  <VTextField v-model="bindPhrase" :label="uid" density="comfortable" :oninput="generateUID"/>
+</template>
